offline/requirements.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
easy-tk==1.0.7.2
future==0.18.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
proglog==0.1.9
pytkgen==1.7
requests==2.24.0
tkintertable==1.3.2
tqdm==4.51.0
urllib3==1.25.11
